WARNING: THIS PULL WILL BREAK CURRENT QUERYEVALUATOR
because ive changed the "boundary" for conditions to pair of pair like relationship parameters instead of just 1 string.
Changed variable types for relationships:
numbers = QueryVar::Stmt by default
"strings" = QueryVar::Procedure for calls and 1st parameter of modifies and uses. For conditions the type varies with the condition attribute type.