Got stunned by and fixed very evil bug in query evaluator, updated modifies test cases
currentFirstVariableName and currentSecondVariableName should be left perfectly alone; removing quotation marks will create a possibility of the name being treated like a synonym of the same name when performing carrtesian, eg Modifies(a, "a"), "a" will be trimmed to a and be treated like the synonym a. Cartesian will very likely return an errorneous "None" result.